+++ sponsoredTransactions/backend/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased changes
 
-- Update `concodium-rust-sdk` dependency and adjust project to be forward-compatible.
+- Update `concodium-rust-sdk` dependency and adjust project to be forward-compatible. Unknown reject reasons when a sponsored transaction is simulated will produce an internal error.
 - Change to use the `contract-client` from the Rust SDK.
 - Add reject reason decoding of reverted transactions during the dry-run.
 - Update submodule link.